fix: make a utils file
