+++ src/typescript.ts
@@ -10,71 +10,13 @@ import { generateMethodParamId, generateMethodResultId } from "@open-rpc/schema-
 import { compile } from "json-schema-to-typescript";
 import { toSafeString } from "json-schema-to-typescript/dist/src/utils";
 import { ContentDescriptorObject, MethodObject, OpenRPC, Schema } from "@open-rpc/meta-schema";
+import { collectAndRefSchemas, getSchemaTypeName, getMethodAliasName } from "./utils";
 
-/**
- * Helper methods
- */
-const collectAndRefSchemas = (schema: Schema): Schema[] => {
-  const newS: Schema = { ...schema };
-  const subS: Schema[][] = [];
-
-  if (schema.anyOf) {
-    subS.push(schema.anyOf);
-    newS.anyOf = schema.anyOf.map(schemaToRef);
-  }
-
-  if (schema.allOf) {
-    subS.push(schema.allOf);
-    newS.allOf = schema.allOf.map(schemaToRef);
-  }
-
-  if (schema.oneOf) {
-    subS.push(schema.oneOf);
-    newS.oneOf = schema.oneOf.map(schemaToRef);
-  }
-
-  if (schema.items) {
-    subS.push(schema.items);
-
-    if (schema.items instanceof Array) {
-      newS.items = schema.items.map(schemaToRef);
-    } else {
-      newS.items = schemaToRef(schema.items);
-    }
-  }
-
-  if (schema.properties) {
-    subS.push(Object.values(schema.properties));
-    newS.properties = _.mapValues(schema.properties, schemaToRef);
-  }
-
-  const subSchemas: Schema[] = _.chain(subS)
-    .flatten()
-    .compact()
-    .value();
-
-  const collectedSubSchemas: Schema[] = _.map(subSchemas, collectAndRefSchemas);
-
-  return _.chain(collectedSubSchemas)
-    .push([newS])
-    .flatten()
-    .value();
-};
-
-const schemaToRef = (s: Schema) => ({ $ref: `#/definitions/${getSchemaTypeName(s)}` });
-
-/**
- * Exported Methods
- */
-export const getMethodAliasName: GetMethodAliasName = ({ name }: MethodObject): string => {
-  return getSchemaTypeName({ title: name, type: "function" });
-};
-
-export const getSchemaTypeName: GetSchemaTypeName = (s: Schema): string => toSafeString(s.title);
 const isComment = (line: string) => {
   const trimmed = line.trim();
   return _.startsWith(trimmed, "/**") || _.startsWith(trimmed, "*") || _.startsWith(trimmed, "*/");
 };
+
 const getDefs = (lines: string) => {
   let commentBuffer: string[] = [];
   return _.chain(lines.split("\n"))
